Fix static issues in AudioD
:Release Notes:
Fix static issues related to Dereference after null check, printf
arguments mismatch and Unchecked return value in AudioD

:Detailed Notes:
Addressed below static issues
CID 8481083 Printf arg type mismatch
CID 8481088 Dereference after null check
CID 8481183 Unchecked return value
CID 8480497 Printf arg type mismatch
CID 8480504 Printf arg type mismatch
CID 8481209 Parse warning

:Testing Performed:
Tested on RPI4 rebased on OSE build #245
Test case: QWA-4099, QWA-4098 QWA-5508
Test Results are attached to PLAT-112248

:QA Notes:
NA

:Issues Addressed:
[PLAT-112248] Fix static issues in AudioD OSE

+++ src/audioMixer.cpp
@@ -75,7 +75,7 @@ utils::vectorVirtualSink AudioMixer::getActiveStreams()
 bool AudioMixer::isStreamActive(EVirtualAudioSink sink)
 {
     PM_LOG_INFO(MSGID_AUDIO_MIXER, INIT_KVCOUNT,\
-        "AudioMixer:: isStreamActive sinkId:%d", sink);
+        "AudioMixer:: isStreamActive sinkId:%d", (int)sink);
     if (IsValidVirtualSink(sink))
     {
         utils::itVirtualSink it;
